Logging: cast enums to int for fmt 12 format calls
fmt 12 (from the Framework spdlog/fmt vendored upgrade) no longer implicitly formats enums. Cast the menu event, district id, and entity type args in the client log/format calls to int, matching prior fmt 8 behavior.

+++ code/client/src/core/hooks/hud.cpp
@@ -12,7 +12,7 @@ typedef int64_t(__fastcall *C_MenuBase__OnScriptedMenuEvent_t)(void *pThis, int,
 C_MenuBase__OnScriptedMenuEvent_t C_MenuBase__OnScriptedMenuEvent_original = nullptr;
 int64_t C_MenuBase__OnScriptedMenuEvent(void *pThis, int unk, SDK::mafia::ui::menu::C_MenuBase::E_ScriptedMenuEvent menuEvent) {
     // TODO: broadcast over network
-    Framework::Logging::GetLogger("Hooks")->debug("C_MenuBase::OnScriptedMenuEvent: unk {} event {}", unk, menuEvent);
+    Framework::Logging::GetLogger("Hooks")->debug("C_MenuBase::OnScriptedMenuEvent: unk {} event {}", unk, static_cast<int>(menuEvent));
     return C_MenuBase__OnScriptedMenuEvent_original(pThis, unk, menuEvent);
 }
